ron-schema: Split TH module into several smaller
